+++ src/Contracts/Factory.php
@@ -9,8 +9,6 @@ interface Factory
      *
      * @param  string|null  $provider
      * @return \Rymanalu\LaravelSimpleUploader\Contracts\Uploader
-     *
-     * @throws \InvalidArgumentException
      */
     public function from($provider = null);
 }