+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## v1.0.4 (2017-01-31)
+
+### Changed
+- Use `$app` instead of `$this->app` when register the Uploader to the Application
+
+## v1.0.3 (2017-01-24)
+
+### Changed
+- Clean up code
+
+## v1.0.2 (2016-12-15)
+
+### Added
+- Added the default file visibility
+
+## v1.0.1 (2016-12-15)
+
+### Added
+- Added contract bindings
+
 ## v1.0.0 (2016-12-11)
 
 ### Added